%  File degreenet/man/sweden.Rd
%  Part of the statnet package, http://statnet.org
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) in
%    http://statnet.org/attribution
%
% Copyright 2003 Mark S. Handcock, University of California-Los Angeles
% Copyright 2007 The statnet Development Team
\name{sweden}
\alias{sweden}
\alias{swedenm}
\alias{swedenf}
\title{Number of sex partners in the last 12 months for men and women in Sweden}
\description{
  This is a data set used in Jones and Handcock (2002) and 
  
  The data are counts of the numbers of sex partners for men
  and women in the last twelve months. The data from the 1996
  ``Sex in Sweden" survey based on a nationwide probability
  sample and financed by the Swedish National Board of Health.
} \usage{
 data(sweden)
}
\source{
We thanks Dr. Bo Lewin, Professor of Sociology, Uppsala
University and head of the research team responsible for the
``Sex in Sweden" study for providing the Swedish data used in
this study.  This research supported by Grant 7R01DA012831-02
from NIDA and Grant 1R01HD041877 from NICHD.
}
\seealso{ayulemle}
\references{
 Lewin, B. (1996).  \emph{Sex in Sweden,}
 Stockholm: National Institute of Public Health.

 Handcock, Mark S. and Jones, James Holland (2004),
 ``Likelihood-Based Inference for Stochastic Models of Sexual Network Formation"
\emph{Theoretical Population Biology},
\url{http://dx.doi.org/10.1016/j.tpb.2003.09.006}.

 Jones, James Holland and Handcock, Mark S. (2003),
\emph{Nature}, 423, 6940, 605-606.

 Handcock, Mark S. and Jones, James Holland (2003),
``An assessment of preferential attachment as a
mechanism for human sexual network formation"
\emph{Proceedings of the Royal Society}, B., 270, 1123-1128.

}
\keyword{datasets}
